+++ src/contentsharecontroller/DefaultContentShareController.ts
@@ -2,8 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import AudioVideoController from '../audiovideocontroller/AudioVideoController';
+import ContentShareObserver from '../contentshareobserver/ContentShareObserver';
+import Maybe from '../maybe/Maybe';
 import MeetingSessionConfiguration from '../meetingsession/MeetingSessionConfiguration';
 import MeetingSessionCredentials from '../meetingsession/MeetingSessionCredentials';
+import AsyncScheduler from '../scheduler/AsyncScheduler';
 import ContentShareConstants from './ContentShareConstants';
 import ContentShareController from './ContentShareController';
 import ContentShareMediaStreamBroker from './ContentShareMediaStreamBroker';
@@ -22,6 +25,8 @@ export default class DefaultContentShareController implements ContentShareContro
       configuration.credentials.joinToken + ContentShareConstants.Modality;
     return contentShareConfiguration;
   }
+
+  private observerQueue: Set<ContentShareObserver> = new Set<ContentShareObserver>();
 
   constructor(
     private mediaStreamBroker: ContentShareMediaStreamBroker,
@@ -42,6 +47,9 @@ export default class DefaultContentShareController implements ContentShareContro
     if (this.mediaStreamBroker.mediaStream.getVideoTracks().length > 0) {
       this.audioVideo.videoTileController.startLocalVideoTile();
     }
+    this.forEachContentShareObserver(observer => {
+      Maybe.of(observer.contentShareDidStart).map(f => f.bind(observer)());
+    });
   }
 
   async startContentShareFromScreenCapture(sourceId?: string): Promise<void> {
@@ -53,14 +61,41 @@ export default class DefaultContentShareController implements ContentShareContro
 
   pauseContentShare(): void {
     this.mediaStreamBroker.toggleMediaStream(false);
+    this.forEachContentShareObserver(observer => {
+      Maybe.of(observer.contentShareDidPause).map(f => f.bind(observer)());
+    });
   }
 
   unpauseContentShare(): void {
     this.mediaStreamBroker.toggleMediaStream(true);
+    this.forEachContentShareObserver(observer => {
+      Maybe.of(observer.contentShareDidUnpause).map(f => f.bind(observer)());
+    });
   }
 
   stopContentShare(): void {
     this.audioVideo.stop();
     this.mediaStreamBroker.cleanup();
+    this.forEachContentShareObserver(observer => {
+      Maybe.of(observer.contentShareDidStop).map(f => f.bind(observer)());
+    });
+  }
+
+  addContentShareObserver(observer: ContentShareObserver): void {
+    this.observerQueue.add(observer);
+  }
+
+  removeContentShareObserver(observer: ContentShareObserver): void {
+    this.observerQueue.delete(observer);
+  }
+
+  forEachContentShareObserver(observerFunc: (observer: ContentShareObserver) => void): void {
+    for (const observer of this.observerQueue) {
+      new AsyncScheduler().start(() => {
+        if (this.observerQueue.has(observer)) {
+          observerFunc(observer);
+        }
+      });
+    }
   }
 }